Create README.md
This programming task is about implementing a web API, so that the same software can be used by
different insurance companies to build their own applications using the provided API. At the
moment, we do not care about the front-end application, which could be any browser or a mobile
app. And we also don’t care about a database that stores the information. In this exercise, we
deal with the API implementation only. You may use a REST client like Postman to quickly see the
results of the API call.
The API is implemented in Python using a package called Flask, which allows you to define HTML
methods GET, POST, PUT etc. Each method returns a JSON object, which can be used by the front-end-application in adequate ways.